Split entity passivation specs into multiple files
